Small fixes, changes to InputManager
Sizes of _players and _controllers in PlayerDetector no longer based on the length of Input.GetJoystickNames since that can also count empty slots. In the same script DetectControllerTypes replaced with DetectControllerType, which only detects the type of the given controller, instead of iterating through the list of them all; is called in DetectPlayers instead of Awake.

In InputManager, changed the values of Joy Num from any controller to either Joystick 1 or Joystick 2.